Refactor(Azure Library): Rename query builder package to 'fluent'
