--- azure-boards-common/src/main/java/com/synopsys/integration/azure/boards/common/service/query/builder/WorkItemQueryWhereJunctionType.java
@@ -1,7 +1,0 @@
-package com.synopsys.integration.azure.boards.common.service.query.builder;
-
-public enum WorkItemQueryWhereJunctionType {
-    AND,
-    OR
-
-}
