+++ azure-boards-common/src/main/java/com/synopsys/integration/azure/boards/common/service/query/fluent/WorkItemQueryOrderBy.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.azure.boards.common.service.query.builder;
+package com.synopsys.integration.azure.boards.common.service.query.fluent;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
