+++ azure-boards-common/src/main/java/com/synopsys/integration/azure/boards/common/service/query/fluent/WorkItemQueryWhereJunctionType.java
@@ -1,0 +1,7 @@
+package com.synopsys.integration.azure.boards.common.service.query.fluent;
+
+public enum WorkItemQueryWhereJunctionType {
+    AND,
+    OR
+
+}
